pyproject.toml
[project]
name = "starfishdata"
version = "0.1.0"
description = "Synthetic data generation cli for starfishdata.ai"
readme = "README.rst"
authors = [
  {name = "Ayush Mudgal", email = "ayushincnca@gmail.com"}
]
maintainers = [
  {name = "Ayush Mudgal", email = "ayushincnca@gmail.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "typer"
]

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.9.0"}
torch = "^2.0.0"
transformers = "^4.35.0"
rich = "^13.5.2"
llama-cpp-python = "^0.2.0" 
huggingface_hub = "^0.19.4"
scikit-learn = "^1.3.0"

[tool.poetry.scripts]
starfishdata = "starfishdata.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


